# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/tests": {
            "post": {
                "tags": [
                    "tests"
                ],
                "summary": "Cerate test",
                "operationId": "testCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "topic",
                                    "course_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "تست بخش دیتابیس"
                                    },
                                    "topic": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "دیتابیس های رابطه ای"
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "test created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "تست بخش دیتابیس"
                                        },
                                        "topic": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "دیتابیس های رابطه ای"
                                        },
                                        "course_id": {
                                            "type": "integer",
                                            "format": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "dateTime",
                                            "example": "2023-01-19T14:09:23.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "dateTime",
                                            "example": "2023-01-19T14:09:23.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tests/{id}": {
            "put": {
                "tags": [
                    "tests"
                ],
                "summary": "Cerate test",
                "operationId": "testUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "test id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "topic",
                                    "course_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "تست بخش دیتابیس"
                                    },
                                    "topic": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "دیتابیس های رابطه ای"
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "test updated succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "درس با موفقیت آپدیت شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tests"
                ],
                "summary": "Delete test",
                "operationId": "testDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "test id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "test updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "درس با موفقیت حذف شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}